+++ Highball/BlogViewController.swift
@@ -15,6 +15,14 @@ class BlogViewController: PostsViewController {
         self.blogName = blogName
         super.init()
         self.navigationItem.title = self.blogName
+
+        let followIcon = FAKIonIcons.iosPersonaddOutlineIconWithSize(30)
+        self.navigationItem.rightBarButtonItem = UIBarButtonItem(
+            image: followIcon.imageWithSize(CGSize(width: 30, height: 30)),
+            style: UIBarButtonItemStyle.Plain,
+            target: self,
+            action: Selector("follow:")
+        )
     }
 
     required init() {
@@ -45,5 +53,49 @@ class BlogViewController: PostsViewController {
     
     override func reblogBlogName() -> (String) {
         return AccountsService.account.blog.name
+    }
+
+    func follow(sender: UIBarButtonItem) {
+        let alertController = UIAlertController(title: nil, message: nil, preferredStyle: UIAlertControllerStyle.ActionSheet)
+
+        alertController.addAction(UIAlertAction(title: "Follow", style: UIAlertActionStyle.Default) { action in
+            TMAPIClient.sharedInstance().follow(self.blogName) { result, error in
+                let alertController = UIAlertController(title: nil, message: nil, preferredStyle: UIAlertControllerStyle.Alert)
+
+                if let error = error {
+                    alertController.title = "Follow Failed"
+                    alertController.message = "Tried to follow \(self.blogName), but failed."
+                } else {
+                    alertController.title = "Followed"
+                    alertController.message = "Successfully followed \(self.blogName)!"
+                }
+
+                alertController.addAction(UIAlertAction(title: "OK", style: UIAlertActionStyle.Default, handler: { _ in }))
+
+                self.presentViewController(alertController, animated: true, completion: nil)
+            }
+        })
+
+        alertController.addAction(UIAlertAction(title: "Unfollow", style: UIAlertActionStyle.Destructive) { action in
+            TMAPIClient.sharedInstance().unfollow(self.blogName) { result, error in
+                let alertController = UIAlertController(title: nil, message: nil, preferredStyle: UIAlertControllerStyle.Alert)
+                
+                if let error = error {
+                    alertController.title = "Unfollow Failed"
+                    alertController.message = "Tried to unfollow \(self.blogName), but failed."
+                } else {
+                    alertController.title = "Unfollowed"
+                    alertController.message = "Successfully unfollowed \(self.blogName)!"
+                }
+                
+                alertController.addAction(UIAlertAction(title: "OK", style: UIAlertActionStyle.Default, handler: { _ in }))
+                
+                self.presentViewController(alertController, animated: true, completion: nil)
+            }
+        })
+
+        alertController.addAction(UIAlertAction(title: "Cancel", style: UIAlertActionStyle.Cancel, handler: { _ in }))
+
+        self.presentViewController(alertController, animated: true, completion: nil)
     }
 }