+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * View posts for a blog.
 * Should finally have fixed a bug that would cause an "end" to the feed.
 * Should actually finally wait for full height calculation when loading posts.
+* Handle asymmetric widths in photoset rows
 
 ## 0.5.3